+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dropinnodex/react
 
+## 0.14.1
+
+### Patch Changes
+
+- Updated dependencies [a2b30ac]
+  - @dropinnodex/client@0.10.0
+
 ## 0.14.0
 
 ### Minor Changes